Add files via upload
Here is an up to date version of the Provisionware client and server.  They are both a part of the same solution on this end, so please ensure that this does not conflict when merging.  I worked exclusively in the Provisionware Client, and did extensive UI fabrication.  NameForm class created to deal with registering name with client session.  Client has a prototype catalog (populated with hard coding in the FormCli constructor).  Cart works, but needs a column to indicate quantity (on client side).  Submit does in fact transmit request to Server.